requirements.txt
asgiref==3.3.1
asn1crypto==1.4.0
attrs==20.3.0
bit==0.7.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coincurve==14.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-requestlogs==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
jsonschema==3.2.0
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
pytz==2020.5
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.0
uvicorn